Color coded neighbourhoods based on number of currently selected business type
